[update] 모집글 CRUD 1. API 수정 2. 빌더 패턴 적용 3. enum class 생성
모집글 기본 CRUD 와 미디어 첨부까지 포스트맨으로 테스트 완료

+++ src/main/java/com/team6/finalproject/post/service/ClubPostServiceImpl.java
@@ -2,6 +2,7 @@ package com.team6.finalproject.post.service;
 
 import com.team6.finalproject.club.entity.Club;
 import com.team6.finalproject.club.repository.ClubRepository;
+import com.team6.finalproject.club.service.ClubService;
 import com.team6.finalproject.common.dto.ApiResponseDto;
 import com.team6.finalproject.common.file.FileUploader;
 import com.team6.finalproject.post.dto.PostRequestDto;
@@ -10,8 +11,8 @@ import com.team6.finalproject.post.entity.Post;
 import com.team6.finalproject.post.repository.PostRepository;
 import com.team6.finalproject.user.entity.User;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,11 +24,12 @@ import java.io.IOException;
 public class ClubPostServiceImpl implements ClubPostService {
 
     private final PostRepository postRepository;
-    private final ClubRepository clubRepository;
+    private final ClubService clubService;
     private final FileUploader fileUploader;
 
     // 모집글 선택 조회
-    @Transactional (readOnly = true)
+    @Override
+    @Transactional(readOnly = true)
     public PostResponseDto getPostById(Long postId) {
         Post post = findPost(postId);
         return new PostResponseDto(post);
@@ -36,12 +38,18 @@ public class ClubPostServiceImpl implements ClubPostService {
     // 모집글 생성
     @Override
     @Transactional
-    public PostResponseDto createdPost(PostRequestDto postRequestDto, Long clubId, User user, MultipartFile multipartFile) throws IOException {
-        Club club = clubRepository.findById(clubId)
-                .orElseThrow(()-> new IllegalArgumentException("존재하지 않는 동호회 입니다."));
+    public PostResponseDto createdPost(PostRequestDto postRequestDto, User user, MultipartFile multipartFile) throws IOException {
+        Club club = clubService.findClub(postRequestDto.getClubId());
+        String media = uploadMedia(multipartFile);
 
-        // 빌더패턴 도입
-        Post post = new Post(postRequestDto, user, club, uploadMedia(multipartFile));
+        Post post = Post.builder()
+                .title(postRequestDto.getTitle())
+                .content(postRequestDto.getContent())
+                .user(user)
+                .club(club)
+                .media(media)
+                .build();
+
         postRepository.save(post);
 
         return new PostResponseDto(post);
@@ -50,9 +58,13 @@ public class ClubPostServiceImpl implements ClubPostService {
     // 모집글 수정
     @Override
     @Transactional
-    public PostResponseDto updatePost(Long postId, PostRequestDto postRequestDto, User user) {
-        Post post = findPost(postId);
+    public PostResponseDto updatePost(Long postId, PostRequestDto postRequestDto, User user, MultipartFile multipartFile) throws IOException {
+        Post post = updateMedia(multipartFile, postId);
+
+        checkedAuthor(post, user);
+
         post.update(postRequestDto);
+
         return new PostResponseDto(post);
     }
 
@@ -61,9 +73,13 @@ public class ClubPostServiceImpl implements ClubPostService {
     @Transactional
     public void deletePost(Long postId, User user) {
         Post post = findPost(postId);
+
+        checkedAuthor(post, user);
+
         post.deletePost();
-        ResponseEntity.ok().body(new ApiResponseDto("모집글 삭제 완료!",200));
+        ResponseEntity.ok().body(new ApiResponseDto("모집글 삭제 완료!", 200));
     }
+
 
     // 모집글에 미디어 업로드
     @Transactional
@@ -73,29 +89,31 @@ public class ClubPostServiceImpl implements ClubPostService {
 
     // 모집글에 미디어 수정
     @Transactional
-    public PostResponseDto updateMedia(MultipartFile file, Long postId) throws IOException {
+    public Post updateMedia(MultipartFile file, Long postId) throws IOException {
         Post post = findPost(postId);
-        if(post.getMedia() != null) {
+        if (post.getMedia() != null) {
             fileUploader.deleteFile(post.getMedia());
         }
 
         String media = fileUploader.upload(file);
         post.updateMedia(media);
-        return new PostResponseDto(post);
+        return post;
     }
 
     // 게시글 존재 여부 확인
     @Override
-    @Transactional
+    @Transactional(readOnly = true)
     public Post findPost(Long postId) {
-        return postRepository.findById(postId).orElseThrow(()->
+        return postRepository.findByActiveId(postId).orElseThrow(() ->
                 new IllegalArgumentException("선택한 글은 존재하지 않습니다.")
         );
     }
+
+    // 글 작성자가 본인인지 확인
+    public void checkedAuthor(Post post, User user) {
+        if (!post.getUser().getUsername().equals(user.getUsername())) {
+            throw new IllegalArgumentException("본인이 작성한 글이 아닙니다.");
+        }
+    }
+
 }
-
-
-// 동호회를 개설한 사람 또는 권한이 부여된 사람이 모집글 작성,수정,삭제를 할 수 있다,,,?
-// 동호회를 개설한 사람이 해당 동호회 멤버에게 권한을 부여 할 수 있다. (다른 동호회 엠버에게 권한 부여 X)
-// 동호회 멤버는 다른사람들에게 권한을 부여할 수 X
-// 그렇다면 이것이 멤버가 구성이 되어야지 가능하다?! -> 동호회 가입을 해야한다?!
